chore: move kubeConfig monitor logic to main thread
