+++ electron/app/createWindow.ts
@@ -1,4 +1,4 @@
-import {BrowserWindow, app, nativeImage} from 'electron';
+import {BrowserWindow, app, ipcMain, nativeImage} from 'electron';
 
 import {indexOf} from 'lodash';
 import {machineIdSync} from 'node-machine-id';
@@ -202,6 +202,11 @@ export const createWindow = (givenPath?: string) => {
     }, NEW_VERSION_CHECK_INTERVAL);
 
     initKubeconfig(dispatch, userHomeDir);
+
+    ipcMain.on('restart-kubeconfig-monitor', (_event, kubeConfigPath) => {
+      initKubeconfig(dispatch, userHomeDir, kubeConfigPath);
+    });
+
     dispatch(setAppRehydrating(false));
 
     const missingDependencies = checkMissingDependencies(APP_DEPENDENCIES);
