+++ src/redux/services/projectConfig.ts
@@ -11,8 +11,6 @@ import {AppConfig, ProjectConfig} from '@models/appconfig';
 
 import {updateProjectConfig} from '@redux/reducers/appConfig';
 
-import {monitorKubeConfig} from './kubeConfigMonitor';
-
 export interface SerializableObject {
   [name: string]: any;
 }
@@ -174,8 +172,6 @@ export const updateProjectSettings = (dispatch: (action: AnyAction) => void, pro
   const projectConfig: ProjectConfig | null = readProjectConfig(projectRootPath);
   if (projectConfig) {
     dispatch(updateProjectConfig({config: projectConfig, fromConfigFile: true}));
-    monitorKubeConfig(dispatch, projectConfig?.kubeConfig?.path);
-
     return;
   }
   dispatch(updateProjectConfig({config: null, fromConfigFile: true}));
